+++ src/include/LSP/Workspace.hpp
@@ -114,9 +114,11 @@ private:
 public:
     std::vector<std::string> getComments(const Luau::ModuleName& moduleName, const Luau::Location& node);
     std::optional<std::string> getDocumentationForType(const Luau::TypeId ty);
+    std::optional<std::string> getDocumentationForTypeReference(const Luau::ModuleName& moduleName, const Luau::ScopePtr& scope,
+        const std::optional<Luau::AstName>& prefix, const Luau::Name& typeName, bool forAutocomplete);
     std::optional<std::string> getDocumentationForAstNode(const Luau::ModuleName& moduleName, const Luau::AstNode* node, const Luau::ScopePtr scope);
     std::optional<std::string> getDocumentationForAutocompleteEntry(const std::string& name, const Luau::AutocompleteEntry& entry,
-        const std::vector<Luau::AstNode*>& ancestry, const Luau::ModulePtr& localModule);
+        const std::vector<Luau::AstNode*>& ancestry, const Luau::ModulePtr& localModule, const Luau::Position& position);
     std::vector<Reference> findAllTableReferences(
         const Luau::TypeId ty, const LSPCancellationToken& cancellationToken, std::optional<Luau::Name> property = std::nullopt);
     std::vector<Reference> findAllFunctionReferences(const Luau::TypeId ty, const LSPCancellationToken& cancellationToken);